requirements.txt
asgiref>=3.8.1
flask-restx>=1.3.0
flask>=3.0.3
langchain-openai>=0.2.6
langchain>=0.3.7
pdfminer-six[image]>=20240706
pydantic>=2.9.2
pymongo[srv]>=4.10.1
pypdf
python-dotenv>=1.0.1
tenacity>=9.0.0
tqdm>=4.67.0
uvicorn>=0.32.0